dags/requirements.txt
fastapi
streamlit
boto3
chromadb
pinecone-client
sentence-transformers
nltk
selenium
docling
mistralai
apache-airflow
apache-airflow-providers-amazon
boto3
